+++ oap-server/server-core/src/main/java/org/apache/skywalking/oap/server/core/CoreModule.java
@@ -68,6 +68,7 @@ import org.apache.skywalking.oap.server.core.server.HTTPHandlerRegister;
 import org.apache.skywalking.oap.server.core.source.SourceReceiver;
 import org.apache.skywalking.oap.server.core.status.ServerStatusService;
 import org.apache.skywalking.oap.server.core.storage.model.IModelManager;
+import org.apache.skywalking.oap.server.core.trace.SpanListenerManager;
 import org.apache.skywalking.oap.server.core.storage.model.ModelCreator;
 import org.apache.skywalking.oap.server.core.storage.model.ModelManipulator;
 import org.apache.skywalking.oap.server.core.worker.IWorkerInstanceGetter;
@@ -93,6 +94,8 @@ public class CoreModule extends ModuleDefine {
         classes.add(NamingControl.class);
         classes.add(IComponentLibraryCatalogService.class);
         classes.add(HierarchyDefinitionService.class);
+
+        classes.add(SpanListenerManager.class);
 
         classes.add(IWorkerInstanceGetter.class);
         classes.add(IWorkerInstanceSetter.class);